+++ src/main/java/com/aiqin/bms/scmp/api/supplier/domain/request/applycontract/dto/ApplyContractDTO.java
@@ -53,7 +53,7 @@ public class ApplyContractDTO extends CommonBean {
     private Byte deliveryCharges;
 
     @ApiModelProperty("卸货费(甲方，乙方承担)")
-    private Long unloadingFee;
+    private BigDecimal unloadingFee;
 
     @ApiModelProperty("其他约定")
     private String otherConventions;
